SBERDOMA-52 Remove scroll logic from MapConstructor
